moves marker management to separate module
